break dependency on usboptions
